add existence check before backing up file

+++ setup_rsync.sh
@@ -3,8 +3,10 @@
 # run this on the raspberry pi 
 # to set up rsync from the client side.
 
-echo "Backing up /etc/rsyncd.conf to /etc/rsyncd.conf.bak"
-mv /etc/rsyncd.conf{,.bak}
+if [ -f /etc/rsyncd.conf ]; then
+	echo "Backing up /etc/rsyncd.conf to /etc/rsyncd.conf.bak"
+	mv /etc/rsyncd.conf{,.bak}
+fi
 
 echo "Installing rsyncd.conf to /etc/rsyncd.conf" 
 cp rsyncd.conf /etc/rsyncd.conf
